feat: proxy support in rancher-cluster
In some dev environments clusters need to communicate via an HTTP proxy. This is added. Additionally, there are scenarios when Rancher needs to be limited in resources (e.g. Vcluster in a Rancher project of the host cluster). Therefore, resources are set. As soon as needed, we can consider setting the resources via Helm values, too. For now, we will keep it hardcoded.